pyproject.toml
[tool.poetry]
name = "businesscontrol"
version = "0.1.0"
description = ""
authors = ["Nikita Smykov <132088678+Apicqq@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.6"
pydantic-settings = "^2.6.1"
sqlalchemy = "^2.0.36"
uvicorn = "^0.32.1"
asyncpg = "^0.30.0"
greenlet = "^3.1.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
bcrypt = "^4.2.1"
email-validator = "^2.2.0"
alembic = "^1.14.0"
python-multipart = "^0.0.19"


[tool.poetry.group.dev.dependencies]
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
